Move the framework's logger out of org.lsposed.lspd, and out of Utils
The last of this project's own code under org.lsposed.lspd, and the third item
34144a8b5 left for later. Twenty-two files reach it, across the legacy API, the
injected framework and the zygisk bridge, all compiling into one artifact — and
nothing in the manager, which wrote its own logger rather than depend on this one.

org.matrix.vector.util, deliberately not one of the three prefixes
daemon/src/main/jni/obfuscation.cpp rewrites. Nothing here is obfuscated today and
this keeps it that way; landing it under .core, .nativebridge or .service would have
started rewriting a class the manager does not compile, for no gain.

Log becomes a top-level class rather than a member of Utils. It held nothing but
static members while being a *non-static inner class*, which Java has accepted only
since release 16, and which meant every call site outside an import wrote Utils.Log
for a type never scoped to an instance of anything. Splitting it also lets the two say
what they are for: Log is the muting-aware stand-in for android.util.Log, and Utils is
the convenience layer that logs under the framework's own tag.

Which methods honour `muted` is uneven — the message-only forms and
w(String, String, Throwable) do, the other Throwable forms do not — and that is
carried over exactly. It looks like an oversight, but changing it changes what a muted
device records, which is a decision about the log rather than about this move.

The tag is the one thing here that must not change: the daemon's log reader routes any
tag beginning "Vector" into its verbose stream, so what these helpers write reaches the
manager's Verbose tab and travels in an exported bug report. It is now documented where
it is declared.

XposedBridge keeps its call fully qualified. That file imports android.util.Log too,
and it is the framework's own that is wanted there — the platform's getStackTraceString
returns an empty string when anything in the cause chain is an UnknownHostException, so
a module logging a failed request landed an empty line.
