+++ legacy/src/main/java/org/matrix/vector/legacy/LegacyDelegateImpl.java
@@ -2,7 +2,7 @@ package org.matrix.vector.legacy;
 
 import android.content.res.XResources;
 
-import org.lsposed.lspd.util.Utils;
+import org.matrix.vector.util.Utils;
 import org.matrix.vector.impl.core.VectorServiceClient;
 import org.matrix.vector.impl.di.LegacyFrameworkDelegate;
 import org.matrix.vector.impl.di.LegacyPackageInfo;
