+++ xposed/src/main/kotlin/org/matrix/vector/impl/core/VectorServiceClient.kt
@@ -5,7 +5,7 @@ import android.os.ParcelFileDescriptor
 import org.matrix.vector.ipc.LoadedModule
 import org.matrix.vector.ipc.IProcessChannel
 import org.matrix.vector.ipc.IFrameworkService
-import org.lsposed.lspd.util.Utils.Log
+import org.matrix.vector.util.Log
 
 /**
  * Singleton client for managing IPC communication with the injected manager service. Handles Binder
