+++ xposed/src/main/kotlin/org/matrix/vector/impl/core/VectorStartup.kt
@@ -4,7 +4,7 @@ import android.app.ActivityThread
 import android.os.Build
 import android.os.IBinder
 import dalvik.system.DexFile
-import org.lsposed.lspd.util.Utils
+import org.matrix.vector.util.Utils
 import org.matrix.vector.ipc.IFrameworkService
 import org.matrix.vector.impl.di.VectorBootstrap
 import org.matrix.vector.impl.hookers.*
